requirements.txt
selenium==4.1.0
pillow==8.3.1
tqdm==4.62.3
webdriver_manager==3.5.2